+++ packages/voice/src/engines/MinimaxEngine.ts
@@ -1,8 +1,8 @@
 import {
-  MINIMAX_GLOBAL_API_URL,
   MINIMAX_CHINA_API_URL,
-  MINIMAX_GLOBAL_VOICE_LIST_URL,
   MINIMAX_CHINA_VOICE_LIST_URL,
+  MINIMAX_GLOBAL_API_URL,
+  MINIMAX_GLOBAL_VOICE_LIST_URL,
 } from '../constants/voiceEngine';
 import { Talk } from '../types/voice';
 import { VoiceEngine } from './VoiceEngine';